pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "oa-mvp"
version = "0.1.0"
description = "OA MVP: FastAPI + SQLite + JWT + vanilla JS"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.29",
  "sqlalchemy>=2.0",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "PyJWT>=2.8",
]

[tool.setuptools.packages.find]
include = ["backend*"]
exclude = ["frontend*"]

[tool.uv]
dev-dependencies = []
